Re-refactor tools exploration
Rename old `action` to `click_tools`
Move all code related to click or the Action classes to click_tools
Re-fix encapsulation
